refactor: Move to Output namespace
